Version and documentation updates
- Upgrade gradle to 5.2.1
- Upgrade Spring Boot to 2.1.3
- Consolidate build.gradle to new format
- Add H2 in place of MySQL as runtime dep
- Update Flyway to v5, switch to runtime dep

+++ src/main/resources/db/migration/schema/V1__initialSchema.sql
@@ -1,10 +1,10 @@
 CREATE TABLE books (
-  id INT not null PRIMARY KEY,
+  id BIGINT not null PRIMARY KEY,
   title VARCHAR(200) not null
 );
 
 CREATE TABLE authors (
-  id int not null PRIMARY KEY,
+  id BIGINT not null PRIMARY KEY,
   first_name VARCHAR(50) not null,
   last_name VARCHAR(50) not null
 )
